Revert "add three bar"
