+++ src/App.js
@@ -1,6 +1,7 @@
+import logo from './logo.svg';
 import './App.css';
 import { Box } from '@mui/material';
-import ChatBody from './maincomp/chatBody';
+import ChatBody from './chat/chatBody';
 
 function App() {
   return (
