--- src/components/chat/Chat.jsx
@@ -1,10 +1,0 @@
-import { Grid } from '@mui/material'
-import React from 'react'
-
-const Chat = () => {
-  return (
-    <Grid sx={{flex:"2"}}>Chat</Grid>
-  )
-}
-
-export default Chat
