--- src/components/details/Details.jsx
@@ -1,10 +1,0 @@
-import { Grid } from '@mui/material'
-import React from 'react'
-
-const Details = () => {
-  return (
-    <Grid sx={{flex:"1"}}>Details</Grid>
-  )
-}
-
-export default Details
