--- src/components/list/chatList/ChatList.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ChatList = () => {
-  return (
-    <div>ChatList</div>
-  )
-}
-
-export default ChatList
